feat: Add Twilio integration and settings
This commit introduces a new Twilio integration, including:

- A new Twilio service for handling API interactions. - A new Twilio
controller to handle voice requests. - New routes for the Twilio
integration. - A new UI settings page for Twilio configuration. -
Updated schema for user and other content-types for assigned_user
relations to one-to-one.

This will allow to integrate phone calls into the CRM.